+++ src/DX12/LLGI.PlatformDX12.cpp
@@ -50,7 +50,7 @@ PlatformDX12::~PlatformDX12()
 {
 	Wait();
 
-	for (int32_t i = 0; i < renderTargets_.size(); i++)
+	for (size_t i = 0; i < renderTargets_.size(); i++)
 	{
 		SafeRelease(renderTargets_[i]);
 	}
